app/core/utils/tradewise_ai.py
import numpy as np
import pandas as pd
import tensorflow as tf
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import LSTM, Dense, Dropout
from sklearn.preprocessing import MinMaxScaler
import talib
from scipy.stats import norm
import gym
from stable_baselines3 import PPO
from datetime import datetime, timedelta

class MarketEnvironment(gym.Env):
    # ... (keep the MarketEnvironment class as is)

class TradewiseAI:
    def __init__(self, lookback_period=100):
        self.lookback_period = lookback_period
        self.scaler = MinMaxScaler()
        self.lstm_model = None
        self.rl_model = None

    def calculate_indicators(self, df):
        df = df.copy()
        
        # Bollinger Bands
        df['bb_middle'], df['bb_upper'], df['bb_lower'] = talib.BBANDS(
            df['close'], 
            timeperiod=20,
            nbdevup=2,
            nbdevdn=2
        )
        
        # RSI and MACD
        df['rsi'] = talib.RSI(df['close'])
        df['macd'], df['macd_signal'], _ = talib.MACD(df['close'])
        
        # Volatility
        df['atr'] = talib.ATR(df['high'], df['low'], df['close'])
        df['volatility'] = df['close'].rolling(window=20).std()
        
        # Price relative to Bollinger Bands
        df['bb_position'] = (df['close'] - df['bb_lower']) / (df['bb_upper'] - df['bb_lower'])
        
        return df

    def build_lstm_model(self, input_shape):
        model = Sequential([
            LSTM(100, return_sequences=True, input_shape=input_shape),
            Dropout(0.2),
            LSTM(50, return_sequences=False),
            Dropout(0.2),
            Dense(25, activation='relu'),
            Dense(1)
        ])
        model.compile(optimizer='adam', loss='huber')
        return model

    def prepare_data(self, df, target_hours=2):
        df_indicators = self.calculate_indicators(df)
        
        features = ['close', 'bb_middle', 'bb_upper', 'bb_lower', 
                   'rsi', 'macd', 'volatility', 'bb_position', 'atr']
        
        scaled_data = self.scaler.fit_transform(df_indicators[features])
        
        X, y = [], []
        for i in range(self.lookback_period, len(scaled_data) - target_hours):
            X.append(scaled_data[i-self.lookback_period:i])
            y.append(scaled_data[i+target_hours, 0])  # Predicting close price
            
        return np.array(X), np.array(y)

    def train_models(self, df):
        X, y = self.prepare_data(df)
        self.lstm_model = self.build_lstm_model((X.shape[1], X.shape[2]))
        self.lstm_model.fit(X, y, epochs=50, batch_size=32, validation_split=0.1)
        
        env = MarketEnvironment(df)
        self.rl_model = PPO('MlpPolicy', env, verbose=0)
        self.rl_model.learn(total_timesteps=10000)

    def generate_forecast(self, current_data, n_scenarios=10):
        df_indicators = self.calculate_indicators(current_data)
        features = ['close', 'bb_middle', 'bb_upper', 'bb_lower', 
                   'rsi', 'macd', 'volatility', 'bb_position', 'atr']
        
        scaled_data = self.scaler.transform(df_indicators[features].tail(self.lookback_period))
        X = scaled_data.reshape(1, self.lookback_period, len(features))
        
        base_pred = self.lstm_model.predict(X)
        
        scenarios = []
        current_price = current_data['close'].iloc[-1]
        
        for _ in range(n_scenarios):
            noise = np.random.normal(0, df_indicators['volatility'].iloc[-1] * 0.1)
            price_pred = self.scaler.inverse_transform(base_pred)[0][0] * (1 + noise)
            
            state = self._get_state(df_indicators.iloc[-1])
            action, _ = self.rl_model.predict(state)
            
            scenario = self._analyze_scenario(
                current_price, 
                price_pred,
                df_indicators.iloc[-1],
                action
            )
            scenarios.append(scenario)
        
        scenarios.sort(key=lambda x: x['expected_return'], reverse=True)
        return scenarios[:10]

    def _get_state(self, row):
        return np.array([
            row['close'], row['bb_middle'], row['bb_upper'], row['bb_lower'],
            row['rsi'], row['macd'], row['volatility'], row['bb_position'],
            row['atr']
        ])

    def _analyze_scenario(self, current_price, predicted_price, indicators, rl_action):
        # ... (keep the _analyze_scenario method as is)

    def generate_trade_suggestions(self, data_feed):
        training_data = data_feed[:-120]  # Use all but last 2 hours for training
        self.train_models(training_data)
        
        current_data = data_feed[-120:]  # Use last 2 hours for current market state
        scenarios = self.generate_forecast(current_data)
        
        suggestions = []
        for i, scenario in enumerate(scenarios, 1):
            suggestion = {
                'Suggestion': f'#{i}',
                'Action': scenario['rl_action'],
                'Summary': {
                    'Current Price': f"${scenario['current_price']:.2f}",
                    'Predicted Price': f"${scenario['predicted_price']:.2f}",
                    'Expected Return': f"{scenario['expected_return']*100:.2f}%",
                    'Confidence': f"{scenario['confidence']*100:.1f}%",
                    'Probability of Profit': f"{scenario['probability_profit']*100:.1f}%"
                },
                'Risk Management': {
                    'Stop Loss': f"${scenario['stop_loss']:.2f} ({(scenario['stop_loss']/scenario['current_price']-1)*100:.1f}%)",
                    'Take Profit': f"${scenario['take_profit']:.2f} ({(scenario['take_profit']/scenario['current_price']-1)*100:.1f}%)",
                    'Risk/Reward Ratio': f"{scenario['risk_reward_ratio']:.2f}",
                    'Suggested Position Size': f"{scenario['suggested_position_size']*100:.1f}% of portfolio"
                },
                'Technical Analysis': {
                    'RSI': f"{scenario['rsi']:.1f}",
                    'Bollinger Position': f"{scenario['bb_position']:.2f}",
                    'Volatility': f"{scenario['volatility']:.2f}",
                    'Signal Strength': {
                        'Overall': f"{scenario['signal_strength']:.2f}",
                        'Bollinger': f"{scenario['technical_indicators']['bb_signal']}",
                        'RSI': f"{scenario['technical_indicators']['rsi_signal']}",
                        'MACD': f"{scenario['technical_indicators']['macd_signal']}"
                    }
                },
                'Forecast Time': scenario['timestamp'].strftime('%Y-%m-%d %H:%M:%S')
            }
            suggestions.append(suggestion)
        
        return suggestions

# You can keep the existing methods in the TradewiseAI class if they're still needed
